test: update tests config
