+++ server/vitest.config.ts
@@ -2,9 +2,15 @@ import { defineConfig } from "vitest/config";
 
 export default defineConfig({
 	test: {
+		pool: "forks",
 		environment: "node",
 		globals: true,
 		setupFiles: ["./tests/setup.ts"],
+		poolOptions: {
+			forks: {
+				singleFork: true,
+			},
+		},
 		coverage: {
 			provider: "v8",
 			//enabled: true,
